feat: added modules/esm implementation
